model-3-predict-image/requirements.txt
Flask
Pillow
tensorflow
numpy
keras
datetime